refactor(utils): move functions around and add comments

+++ src/window_utils.h
@@ -1,9 +1,7 @@
 #pragma once
 
-#include <stddef.h>
-
 int get_integer_scale();
 int get_texture_origin_x();
 int get_texture_origin_y();
-size_t get_mouse_x();
-size_t get_mouse_y();
+int get_mouse_x();
+int get_mouse_y();
